# contracts/generated/openapi/modules/buying.yaml
openapi: 3.1.0
info:
  title: Letron ERPNext Buying API
  version: 0.1.0
servers:
- url: http://127.0.0.1:8080
...
security:
- frappeToken: []
paths:
  /api/v1/buying/purchase-orders:
    get:
      tags:
      - 'Module: Buying'
      summary: List Purchase Order records
      description: List typed Purchase Order documents.
      operationId: listPurchaseOrder
      parameters:
      - &id002
        $ref: '#/components/parameters/RequestId'
      - &id015
        name: fields
        in: query
        schema:
          type: string
        description: JSON field-name array
      - &id016
        name: filters
        in: query
        schema:
          type: string
        description: JSON Frappe filters
      - &id017
        name: order_by
        in: query
        schema:
          type: string
      - &id018
        name: limit_page_length
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 1000
      - &id019
        name: limit_start
        in: query
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: Typed resource list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: &id003
                      $ref: '#/components/schemas/PurchaseOrder'
                  message: {}
                required:
                - data
        '400': &id004
          description: Invalid request
          content: &id001
            application/json:
              schema:
                $ref: '#/components/schemas/FrappeError'
        '401': &id005
          description: Authentication required
          content: *id001
        '403': &id006
          description: Permission denied
          content: *id001
        '404': &id007
          description: Resource not found
          content: *id001
        '409': &id008
          description: Idempotency or optimistic concurrency conflict
          content: *id001
        '417': &id009
          description: Native Frappe validation failure
          content: *id001
        '429': &id010
          description: Rate limited
          content: *id001
        '500': &id011
          description: Server or rollback failure
          content: *id001
      x-test-status: passed
      x-test-level: docker-runtime
      x-test-evidence:
        suite: extended-api
        test: tests/integration/test_extended_api_runtime.py::test_extended_auth_system_resources_and_business_flows
    post:
      tags:
      - 'Module: Buying'
      summary: Create Purchase Order
      description: Create a Purchase Order using the runtime DocType schema.
      operationId: createPurchaseOrder
      parameters: &id020
      - *id002
      - &id012
        $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: &id013
              type: object
              properties:
                name: &id030
                  type: string
                supplier_section: &id031
                  type: string
                  x-frappe-options: fa fa-user
                naming_series: &id032
                  type: string
                  enum:
                  - PUR-ORD-.YYYY.-
                supplier: &id033
                  type: string
                  x-frappe-target-doctype: Supplier
                company: &id034
                  type: string
                  x-frappe-target-doctype: Company
                column_break1: &id035
                  type: string
                transaction_date: &id036
                  type: string
                  format: date
                schedule_date: &id037
                  type: string
                  format: date
                order_confirmation_no: &id038
                  type: string
                order_confirmation_date: &id039
                  type: string
                  format: date
                drop_ship: &id040
                  type: string
                column_break_19: &id041
                  type: string
                customer_contact_person: &id042
                  type: string
                  x-frappe-target-doctype: Contact
                customer_contact_display: &id043
                  type: string
                customer_contact_mobile: &id044
                  type: string
                  x-frappe-options: Phone
                  writeOnly: true
                customer_contact_email: &id045
                  type: string
                  x-frappe-options: Email
                  writeOnly: true
                section_addresses: &id046
                  type: string
                supplier_address: &id047
                  type: string
                  x-frappe-target-doctype: Address
                contact_person: &id048
                  type: string
                  x-frappe-target-doctype: Contact
                col_break_address: &id049
                  type: string
                shipping_address: &id050
                  type: string
                  x-frappe-target-doctype: Address
                currency_and_price_list: &id051
                  type: string
                  x-frappe-options: fa fa-tag
                currency: &id052
                  type: string
                  x-frappe-target-doctype: Currency
                conversion_rate: &id053
                  type: number
                cb_price_list: &id054
                  type: string
                buying_price_list: &id055
                  type: string
                  x-frappe-target-doctype: Price List
                plc_conversion_rate: &id056
                  type: number
                ignore_pricing_rule: &id057
                  type: boolean
                set_warehouse: &id058
                  type: string
                  x-frappe-target-doctype: Warehouse
                is_subcontracted: &id059
                  type: boolean
                supplier_warehouse: &id060
                  type: string
                  x-frappe-target-doctype: Warehouse
                items_section: &id061
                  type: string
                  x-frappe-options: fa fa-shopping-cart
                scan_barcode: &id062
                  type: string
                  x-frappe-options: Barcode
                items: &id063
                  type: array
                  items:
                    $ref: '#/components/schemas/PurchaseOrderItem'
                section_break_48: &id064
                  type: string
                raw_material_details: &id065
                  type: string
                sb_last_purchase: &id066
                  type: string
                column_break_26: &id067
                  type: string
                taxes_section: &id068
                  type: string
                  x-frappe-options: fa fa-money
                taxes_and_charges: &id069
                  type: string
                  x-frappe-target-doctype: Purchase Taxes and Charges Template
                column_break_50: &id070
                  type: string
                shipping_rule: &id071
                  type: string
                  x-frappe-target-doctype: Shipping Rule
                section_break_52: &id072
                  type: string
                taxes: &id073
                  type: array
                  items:
                    $ref: '#/components/schemas/PurchaseTaxesandCharges'
                sec_tax_breakup: &id074
                  type: string
                totals: &id075
                  type: string
                  x-frappe-options: fa fa-money
                column_break_39: &id076
                  type: string
                discount_section: &id077
                  type: string
                apply_discount_on: &id078
                  type: string
                  enum:
                  - Grand Total
                  - Net Total
                column_break_45: &id079
                  type: string
                additional_discount_percentage: &id080
                  type: number
                discount_amount: &id081
                  type: number
                  x-frappe-options: currency
                totals_section: &id082
                  type: string
                column_break4: &id083
                  type: string
                disable_rounded_total: &id084
                  type: boolean
                payment_schedule_section: &id085
                  type: string
                payment_terms_template: &id086
                  type: string
                  x-frappe-target-doctype: Payment Terms Template
                payment_schedule: &id087
                  type: array
                  items:
                    $ref: '#/components/schemas/PaymentSchedule'
                terms_section_break: &id088
                  type: string
                  x-frappe-options: fa fa-legal
                tc_name: &id089
                  type: string
                  x-frappe-target-doctype: Terms and Conditions
                terms: &id090
                  type: string
                column_break_74: &id091
                  type: string
                column_break5: &id092
                  type: string
                letter_head: &id093
                  type: string
                  x-frappe-target-doctype: Letter Head
                select_print_heading: &id094
                  type: string
                  x-frappe-target-doctype: Print Heading
                column_break_86: &id095
                  type: string
                group_same_items: &id096
                  type: boolean
                language: &id097
                  type: string
                from_date: &id098
                  type: string
                  format: date
                to_date: &id099
                  type: string
                  format: date
                column_break_97: &id100
                  type: string
                update_auto_repeat_reference: &id101
                  type: string
                tax_category: &id102
                  type: string
                  x-frappe-target-doctype: Tax Category
                set_reserve_warehouse: &id103
                  type: string
                  x-frappe-target-doctype: Warehouse
                tracking_section: &id104
                  type: string
                column_break_75: &id105
                  type: string
                billing_address: &id106
                  type: string
                  x-frappe-target-doctype: Address
                before_items_section: &id107
                  type: string
                items_col_break: &id108
                  type: string
                accounting_dimensions_section: &id109
                  type: string
                cost_center: &id110
                  type: string
                  x-frappe-target-doctype: Cost Center
                dimension_col_break: &id111
                  type: string
                project: &id112
                  type: string
                  x-frappe-target-doctype: Project
                set_from_warehouse: &id113
                  type: string
                  x-frappe-target-doctype: Warehouse
                terms_tab: &id114
                  type: string
                more_info_tab: &id115
                  type: string
                column_break_7: &id116
                  type: string
                column_break_40: &id117
                  type: string
                column_break_53: &id118
                  type: string
                address_and_contact_tab: &id119
                  type: string
                company_billing_address_section: &id120
                  type: string
                additional_info_section: &id121
                  type: string
                column_break_99: &id122
                  type: string
                column_break_103: &id123
                  type: string
                incoterm: &id124
                  type: string
                  x-frappe-target-doctype: Incoterm
                named_place: &id125
                  type: string
                shipping_address_section: &id126
                  type: string
                connections_tab: &id127
                  type: string
                advance_payment_status: &id128
                  type: string
                  enum:
                  - Not Initiated
                  - Initiated
                  - Partially Paid
                  - Fully Paid
                  writeOnly: true
                has_unit_price_items: &id129
                  type: boolean
                  writeOnly: true
                dispatch_address: &id130
                  type: string
                  x-frappe-target-doctype: Address
                supplier_group: &id131
                  type: string
                  x-frappe-target-doctype: Supplier Group
                  writeOnly: true
                last_scanned_warehouse: &id132
                  type: string
                item_wise_tax_details: &id133
                  type: array
                  items:
                    $ref: '#/components/schemas/ItemWiseTaxDetail'
                  writeOnly: true
                column_break_jkoz: &id134
                  type: string
                base_totals_section: &id135
                  type: string
                  x-frappe-options: Company:company:default_currency
                section_break_tnkm: &id136
                  type: string
                transaction_time: &id137
                  type: string
                  format: time
                auto_repeat_section: &id138
                  type: string
                title: &id139
                  type: string
              required:
              - company
              - conversion_rate
              - currency
              - items
              - name
              - naming_series
              - supplier
              - transaction_date
              additionalProperties: true
              x-module: Buying
            example:
              naming_series: ''
              supplier: ''
              company: ''
              transaction_date: ''
              currency: ''
              conversion_rate: 0
              items: []
      responses:
        '200':
          description: Created typed resource
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: *id003
                required:
                - data
        '400': *id004
        '401': *id005
        '403': *id006
        '404': *id007
        '409': *id008
        '417': *id009
        '429': *id010
        '500': *id011
      x-test-status: passed
      x-test-level: docker-runtime
      x-test-evidence:
        suite: extended-api
        test: tests/integration/test_extended_api_runtime.py::test_extended_auth_system_resources_and_business_flows
...
  /api/v1/crm/request-for-quotations:
    get:
      tags:
      - 'Module: Buying'
      summary: List Request for Quotation records
      description: List typed Request for Quotation documents.
      operationId: listRequestforQuotation
      parameters:
      - *id002
      - *id015
      - *id016
      - *id017
      - *id018
      - *id019
      responses:
        '200':
          description: Typed resource list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: &id021
                      $ref: '#/components/schemas/RequestforQuotation'
                  message: {}
                required:
                - data
        '400': *id004
        '401': *id005
        '403': *id006
        '404': *id007
        '409': *id008
        '417': *id009
        '429': *id010
        '500': *id011
      x-test-status: passed
      x-test-level: docker-runtime
      x-test-evidence:
        suite: crm-preorder
        test: tests/integration/test_crm_preorder.py::test_crm_preorder_contract_and_lifecycle
    post:
      tags:
      - 'Module: Buying'
      summary: Create Request for Quotation
      description: Create a Request for Quotation using the runtime DocType schema.
      operationId: createRequestforQuotation
      parameters: *id020
      requestBody:
        required: true
        content:
          application/json:
            schema: &id022
              type: object
              properties:
                name: &id140
                  type: string
                naming_series: &id141
                  type: string
                  enum:
                  - PUR-RFQ-.YYYY.-
                company: &id142
                  type: string
                  x-frappe-target-doctype: Company
                column_break1: &id143
                  type: string
                transaction_date: &id144
                  type: string
                  format: date
                suppliers_section: &id145
                  type: string
                suppliers: &id146
                  type: array
                  items:
                    $ref: '#/components/schemas/RequestforQuotationSupplier'
                items_section: &id147
                  type: string
                  x-frappe-options: fa fa-shopping-cart
                items: &id148
                  type: array
                  items:
                    $ref: '#/components/schemas/RequestforQuotationItem'
                supplier_response_section: &id149
                  type: string
                email_template: &id150
                  type: string
                  x-frappe-target-doctype: Email Template
                message_for_supplier: &id151
                  type: string
                terms_section_break: &id152
                  type: string
                  x-frappe-options: fa fa-legal
                tc_name: &id153
                  type: string
                  x-frappe-target-doctype: Terms and Conditions
                terms: &id154
                  type: string
                printing_settings: &id155
                  type: string
                select_print_heading: &id156
                  type: string
                  x-frappe-target-doctype: Print Heading
                letter_head: &id157
                  type: string
                  x-frappe-target-doctype: Letter Head
                more_info: &id158
                  type: string
                  x-frappe-options: fa fa-file-text
                col_break_email_1: &id159
                  type: string
                preview: &id160
                  type: string
                sec_break_email_2: &id161
                  type: string
                schedule_date: &id162
                  type: string
                  format: date
                incoterm: &id163
                  type: string
                  x-frappe-target-doctype: Incoterm
                named_place: &id164
                  type: string
                send_attached_files: &id165
                  type: boolean
                send_document_print: &id166
                  type: boolean
                billing_address: &id167
                  type: string
                  x-frappe-target-doctype: Address
                has_unit_price_items: &id168
                  type: boolean
                  writeOnly: true
                subject: &id169
                  type: string
                mfs_html: &id170
                  type: string
                use_html: &id171
                  type: boolean
                  writeOnly: true
                address_and_contact_tab: &id172
                  type: string
                column_break_czul: &id173
                  type: string
                shipping_address: &id174
                  type: string
                  x-frappe-target-doctype: Address
                title: &id175
                  type: string
              required:
              - company
              - items
              - name
              - naming_series
              - subject
              - suppliers
              - transaction_date
              additionalProperties: true
              x-module: Buying
            example:
              naming_series: ''
              company: ''
              transaction_date: ''
              suppliers: []
              items: []
              subject: ''
      responses:
        '200':
          description: Created typed resource
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: *id021
                required:
                - data
        '400': *id004
        '401': *id005
        '403': *id006
        '404': *id007
        '409': *id008
        '417': *id009
        '429': *id010
        '500': *id011
      x-test-status: passed
      x-test-level: docker-runtime
      x-test-evidence:
        suite: crm-preorder
        test: tests/integration/test_crm_preorder.py::test_crm_preorder_contract_and_lifecycle
...
components:
  securitySchemes:
    frappeToken:
      type: apiKey
      in: header
      name: Authorization
      description: 'Frappe token authentication: token api_key:api_secret'
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    ItemWiseTaxDetail:
      type: object
      properties:
        name:
          type: string
        item_row:
          type: string
        tax_row:
          type: string
        rate:
          type: number
        amount:
          type: number
          x-frappe-options: Company:company:default_currency
        taxable_amount:
          type: number
          x-frappe-options: Company:company:default_currency
      required:
      - item_row
      - name
      - tax_row
      additionalProperties: true
      x-module: Accounts
...
    PurchaseOrder:
      type: object
      properties:
        name: *id030
        supplier_section: *id031
        naming_series: *id032
        supplier: *id033
        supplier_name:
          type: string
          readOnly: true
        company: *id034
        column_break1: *id035
        transaction_date: *id036
        schedule_date: *id037
        order_confirmation_no: *id038
        order_confirmation_date: *id039
        amended_from:
          type: string
          x-frappe-target-doctype: Purchase Order
          readOnly: true
        drop_ship: *id040
        customer:
          type: string
          x-frappe-target-doctype: Customer
          readOnly: true
        customer_name:
          type: string
          readOnly: true
        column_break_19: *id041
        customer_contact_person: *id042
        customer_contact_display: *id043
        customer_contact_mobile: *id044
        customer_contact_email: *id045
        section_addresses: *id046
        supplier_address: *id047
        contact_person: *id048
        address_display:
          type: string
          readOnly: true
        contact_display:
          type: string
          readOnly: true
        contact_mobile:
          type: string
          x-frappe-options: Phone
          readOnly: true
        contact_email:
          type: string
          x-frappe-options: Email
          readOnly: true
        col_break_address: *id049
        shipping_address: *id050
        shipping_address_display:
          type: string
          readOnly: true
        currency_and_price_list: *id051
        currency: *id052
        conversion_rate: *id053
        cb_price_list: *id054
        buying_price_list: *id055
        price_list_currency:
          type: string
          x-frappe-target-doctype: Currency
          readOnly: true
        plc_conversion_rate: *id056
        ignore_pricing_rule: *id057
        set_warehouse: *id058
        is_subcontracted: *id059
        supplier_warehouse: *id060
        items_section: *id061
        scan_barcode: *id062
        items: *id063
        section_break_48: *id064
        pricing_rules:
          type: array
          items:
            $ref: '#/components/schemas/PricingRuleDetail'
          readOnly: true
        raw_material_details: *id065
        supplied_items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseOrderItemSupplied'
          readOnly: true
        sb_last_purchase: *id066
        total_qty:
          type: number
          readOnly: true
        base_total:
          type: number
          x-frappe-options: Company:company:default_currency
          readOnly: true
        base_net_total:
          type: number
          x-frappe-options: Company:company:default_currency
          readOnly: true
        column_break_26: *id067
        total:
          type: number
          x-frappe-options: currency
          readOnly: true
        net_total:
          type: number
          x-frappe-options: currency
          readOnly: true
        total_net_weight:
          type: number
          readOnly: true
        taxes_section: *id068
        taxes_and_charges: *id069
        column_break_50: *id070
        shipping_rule: *id071
        section_break_52: *id072
        taxes: *id073
        sec_tax_breakup: *id074
        other_charges_calculation:
          type: string
          readOnly: true
        totals: *id075
        base_taxes_and_charges_added:
          type: number
          x-frappe-options: Company:company:default_currency
          readOnly: true
        base_taxes_and_charges_deducted:
          type: number
          x-frappe-options: Company:company:default_currency
          readOnly: true
        base_total_taxes_and_charges:
          type: number
          x-frappe-options: Company:company:default_currency
          readOnly: true
        column_break_39: *id076
        taxes_and_charges_added:
          type: number
          x-frappe-options: currency
          readOnly: true
        taxes_and_charges_deducted:
          type: number
          x-frappe-options: currency
          readOnly: true
        total_taxes_and_charges:
          type: number
          x-frappe-options: currency
          readOnly: true
        discount_section: *id077
        apply_discount_on: *id078
        base_discount_amount:
          type: number
          x-frappe-options: Company:company:default_currency
          readOnly: true
        column_break_45: *id079
        additional_discount_percentage: *id080
        discount_amount: *id081
        totals_section: *id082
        base_grand_total:
          type: number
          x-frappe-options: Company:company:default_currency
          readOnly: true
        base_rounding_adjustment:
          type: number
          x-frappe-options: Company:company:default_currency
          readOnly: true
        base_in_words:
          type: string
          readOnly: true
        base_rounded_total:
          type: number
          x-frappe-options: Company:company:default_currency
          readOnly: true
        column_break4: *id083
        grand_total:
          type: number
          x-frappe-options: currency
          readOnly: true
        rounding_adjustment:
          type: number
          x-frappe-options: currency
          readOnly: true
        rounded_total:
          type: number
          x-frappe-options: currency
          readOnly: true
        disable_rounded_total: *id084
        in_words:
          type: string
          readOnly: true
        advance_paid:
          type: number
          x-frappe-options: party_account_currency
          readOnly: true
        payment_schedule_section: *id085
        payment_terms_template: *id086
        payment_schedule: *id087
        terms_section_break: *id088
        tc_name: *id089
        terms: *id090
        status:
          type: string
          enum:
          - Draft
          - On Hold
          - To Receive and Bill
          - To Bill
          - To Receive
          - Completed
          - Cancelled
          - Closed
          - Delivered
          readOnly: true
        ref_sq:
          type: string
          x-frappe-target-doctype: Supplier Quotation
          readOnly: true
        party_account_currency:
          type: string
          x-frappe-target-doctype: Currency
          readOnly: true
          writeOnly: true
        inter_company_order_reference:
          type: string
          x-frappe-target-doctype: Sales Order
          readOnly: true
        column_break_74: *id091
        per_received:
          type: number
          readOnly: true
        per_billed:
          type: number
          readOnly: true
        column_break5: *id092
        letter_head: *id093
        select_print_heading: *id094
        column_break_86: *id095
        group_same_items: *id096
        language: *id097
        from_date: *id098
        to_date: *id099
        column_break_97: *id100
        auto_repeat:
          type: string
          x-frappe-target-doctype: Auto Repeat
          readOnly: true
        update_auto_repeat_reference: *id101
        tax_category: *id102
        set_reserve_warehouse: *id103
        tracking_section: *id104
        column_break_75: *id105
        billing_address: *id106
        billing_address_display:
          type: string
          readOnly: true
        before_items_section: *id107
        items_col_break: *id108
        is_internal_supplier:
          type: boolean
          readOnly: true
        represents_company:
          type: string
          x-frappe-target-doctype: Company
          readOnly: true
        accounting_dimensions_section: *id109
        cost_center: *id110
        dimension_col_break: *id111
        project: *id112
        is_old_subcontracting_flow:
          type: boolean
          readOnly: true
          writeOnly: true
        set_from_warehouse: *id113
        terms_tab: *id114
        more_info_tab: *id115
        column_break_7: *id116
        column_break_40: *id117
        column_break_53: *id118
        address_and_contact_tab: *id119
        company_billing_address_section: *id120
        additional_info_section: *id121
        column_break_99: *id122
        column_break_103: *id123
        incoterm: *id124
        named_place: *id125
        shipping_address_section: *id126
        connections_tab: *id127
        advance_payment_status: *id128
        has_unit_price_items: *id129
        dispatch_address: *id130
        dispatch_address_display:
          type: string
          readOnly: true
        supplier_group: *id131
        last_scanned_warehouse: *id132
        mps:
          type: string
          x-frappe-target-doctype: Master Production Schedule
          readOnly: true
        item_wise_tax_details: *id133
        column_break_jkoz: *id134
        base_totals_section: *id135
        section_break_tnkm: *id136
        transaction_time: *id137
        auto_repeat_section: *id138
        title: *id139
      required:
      - company
      - conversion_rate
      - currency
      - items
      - name
      - naming_series
      - status
      - supplier
      - transaction_date
      additionalProperties: true
      x-module: Buying
...
    FrappeError:
      type: object
      properties:
        exc_type:
          type: string
        exception:
          type: string
        _server_messages:
          type: string
        request_id:
          type: string
...